Task 2 and task 3 added to the code
Extension: (if you complete tasks 1 and 2, the extension task is to make this application interactive)
Adapt the application so that it will accept commands from a user, and show the user what they want depending on what command they use. Allow the user to enter the commands
Books - to list books recommended by the book club
Members - to list the members
Allow users to add new members
Allow users to add new books to the recommended books list
If you have done all this and want to try one more thing, add a command for the user to add new books to a member’s reading list.